requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
langchain-text-splitters
langchain-openai
chromadb
langchain-community
sentence_transformers
langchain_huggingface
faiss-cpu
pylance
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
